backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.1
bleach==6.0.0
blinker==1.6.2
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
crispy-tailwind==0.5.0
cron-descriptor==1.4.5
cryptography==41.0.3
Cython==3.0.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==4.2.4
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crispy-forms==2.0
django-filter==24.3
django-import-export==4.1.1
django-phonenumber-field==7.2.0
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.3
fqdn==1.5.1
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kombu==5.4.2
Markdown==3.7
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
MouseInfo==0.1.3
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook_shim==0.2.3
numpy==1.25.2
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.26
pickleshare==0.7.5
Pillow==10.1.0
pkg_resources==0.0.0
platformdirs==3.10.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.18.0
pycryptodome-test-vectors==1.0.12
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==7.4.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-memcached==1.62
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.2.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==0.5.1
rpds-py==0.10.2
scikit-learn==1.3.0
scipy==1.11.2
semver==2.7.9
Send2Trash==1.8.2
sha3==0.2.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==1.5.7
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==1.33
uri-template==1.3.0
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
wikipedia==1.4.0
yarl==1.9.4
zipp==3.16.2
